Refactor effects to audioreactive, generative and common effects
