+++ configs/defence.json
@@ -1,1 +1,1 @@
-{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "a903c28fdd004c068c3c97bbb6fe21a5", "to_node_channel": 0, "to_node_uid": "9e8a2b8a01434826bae425d10357ba9e", "uid": "d592ecfb68ab4c039077c9c33b1356d2"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "192.168.9.241", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "9e8a2b8a01434826bae425d10357ba9e"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.DefenceMode", "py/state": {"num_pixels": 300, "scale": 0.2}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "a903c28fdd004c068c3c97bbb6fe21a5"}}], "recordTimings": true}}
+{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "2d1722c7987e4a1e90b6285a3eaadba6", "to_node_channel": 0, "to_node_uid": "b461c6c18432495196be000695579e51", "uid": "e8a1134c68cb4a688adab52e455bcc92"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "127.0.0.1", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "b461c6c18432495196be000695579e51"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.generative.DefenceMode", "py/state": {"num_pixels": 300, "scale": 0.2}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "2d1722c7987e4a1e90b6285a3eaadba6"}}], "recordTimings": true}}
