+++ configs/movingLight.json
@@ -1,1 +1,1 @@
-{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "dc4bbb67957245d6b8b3e1fbbf1a5337", "to_node_channel": 0, "to_node_uid": "af6857083bb24a41a68695750ee5fc22", "uid": "983fccc370c24254b59e564c1ae36a34"}, {"from_node_channel": 0, "from_node_uid": "0c62bae995c847ecac5bed8ec65a47ce", "to_node_channel": 1, "to_node_uid": "af6857083bb24a41a68695750ee5fc22", "uid": "4811113ab1094798b5fde30c59f92fcf"}, {"from_node_channel": 0, "from_node_uid": "af6857083bb24a41a68695750ee5fc22", "to_node_channel": 0, "to_node_uid": "847000dd8f2d42baa55733b8669aea0c", "uid": "60c6f0bb5fa14e36ae11fde8bea66f3c"}, {"from_node_channel": 0, "from_node_uid": "847000dd8f2d42baa55733b8669aea0c", "to_node_channel": 0, "to_node_uid": "cce5cff768b042ed93fc1fb41a7dffd2", "uid": "7b61a575a0944e02bf73a6bedb25c10b"}, {"from_node_channel": 0, "from_node_uid": "847000dd8f2d42baa55733b8669aea0c", "to_node_channel": 1, "to_node_uid": "cce5cff768b042ed93fc1fb41a7dffd2", "uid": "0bb1143ad55c4daf91ce260cc3f825ec"}, {"from_node_channel": 0, "from_node_uid": "cce5cff768b042ed93fc1fb41a7dffd2", "to_node_channel": 0, "to_node_uid": "9bc01df9422541c2b0014d0d30cbc679", "uid": "6839924e25cd4f03b90c7bb4d0847ba6"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.audio.AudioInput", "py/state": {"chunk_rate": 60, "device_index": null, "num_channels": 2}}, "numInputChannels": 0, "numOutputChannels": 2, "uid": "dc4bbb67957245d6b8b3e1fbbf1a5337"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "192.168.9.241", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "9bc01df9422541c2b0014d0d30cbc679"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.ColorWheelEffect", "py/state": {"cycle_time": 30.0, "luminocity": 0.5, "num_pixels": 150, "offset": 0.0, "saturation": 1.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "0c62bae995c847ecac5bed8ec65a47ce"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.MovingLightEffect", "py/state": {"dim_time": 2.0, "fs": 48000, "highcut_hz": 300.0, "lowcut_hz": 50.0, "num_pixels": 150, "speed": 100.0}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "af6857083bb24a41a68695750ee5fc22"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.MirrorEffect", "py/state": {"mirror_lower": true, "num_pixels": 150, "recursion": 0}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "cd930e6a91cd4c5b90f73b58abf2a40c"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.AfterGlowEffect", "py/state": {"glow_time": 1.0}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "847000dd8f2d42baa55733b8669aea0c"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.Append", "py/state": {"flip0": true, "flip1": false, "flip2": false, "flip3": false, "flip4": false, "flip5": false, "flip6": false, "flip7": false, "num_channels": 2}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "cce5cff768b042ed93fc1fb41a7dffd2"}}], "recordTimings": true}}
+{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "4599ddc058304fb588f83265d41a0e26", "to_node_channel": 0, "to_node_uid": "cddbaf13fd754e1c94ac684588073b8d", "uid": "1106e3ca2a114cda96e80ebfe334d8f8"}, {"from_node_channel": 0, "from_node_uid": "e355ea30ba534b02b30c167eb79c95c3", "to_node_channel": 1, "to_node_uid": "cddbaf13fd754e1c94ac684588073b8d", "uid": "981d3bfd8b094b7db132881f240e818b"}, {"from_node_channel": 0, "from_node_uid": "cddbaf13fd754e1c94ac684588073b8d", "to_node_channel": 0, "to_node_uid": "bfd31046b87945b992779d17ffa4fbd0", "uid": "1866c464e0a543909fd47b57ff351b5c"}, {"from_node_channel": 0, "from_node_uid": "bfd31046b87945b992779d17ffa4fbd0", "to_node_channel": 0, "to_node_uid": "789b5629e1814c17b8fc2f2857f54be2", "uid": "40b6e15657754bb395f9705584244cf7"}, {"from_node_channel": 0, "from_node_uid": "bfd31046b87945b992779d17ffa4fbd0", "to_node_channel": 1, "to_node_uid": "789b5629e1814c17b8fc2f2857f54be2", "uid": "ba76e860e1b54e8ab556328045e4239f"}, {"from_node_channel": 0, "from_node_uid": "789b5629e1814c17b8fc2f2857f54be2", "to_node_channel": 0, "to_node_uid": "393dd5810706434f9d637c0aac347cbf", "uid": "9716975035ea4b64bda004bb78cff005"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.audio.AudioInput", "py/state": {"chunk_rate": 60, "device_index": null, "num_channels": 2}}, "numInputChannels": 0, "numOutputChannels": 2, "uid": "4599ddc058304fb588f83265d41a0e26"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "127.0.0.1", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "393dd5810706434f9d637c0aac347cbf"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.ColorWheelEffect", "py/state": {"cycle_time": 30.0, "luminocity": 0.5, "num_pixels": 150, "offset": 0.0, "saturation": 1.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "e355ea30ba534b02b30c167eb79c95c3"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.audioreactive.MovingLightEffect", "py/state": {"dim_time": 2.0, "fs": 44100, "highcut_hz": 300.0, "lowcut_hz": 50.0, "num_pixels": 150, "speed": 100.0}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "cddbaf13fd754e1c94ac684588073b8d"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.MirrorEffect", "py/state": {"mirror_lower": true, "num_pixels": 150, "recursion": 0}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "2108b33418734df8950d176d69676681"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.AfterGlowEffect", "py/state": {"glow_time": 1.0}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "bfd31046b87945b992779d17ffa4fbd0"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.Append", "py/state": {"flip0": true, "flip1": false, "flip2": false, "flip3": false, "flip4": false, "flip5": false, "flip6": false, "flip7": false, "num_channels": 2}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "789b5629e1814c17b8fc2f2857f54be2"}}], "recordTimings": true}}
