+++ configs/spectrum.json
@@ -1,1 +1,1 @@
-{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "341b261ea3e546aa84152dea2e2b86d3", "to_node_channel": 0, "to_node_uid": "22424635fb4d4c1685225c044664770e", "uid": "c0f0e7f0e93a454fad91400c88a67f51"}, {"from_node_channel": 0, "from_node_uid": "57ad95eb504b4116875dca620ef0d6b3", "to_node_channel": 1, "to_node_uid": "22424635fb4d4c1685225c044664770e", "uid": "528f8111a5104e3aa4d6585653554543"}, {"from_node_channel": 0, "from_node_uid": "1b660e63dc2e4556b8ec958ff1ce2c64", "to_node_channel": 2, "to_node_uid": "22424635fb4d4c1685225c044664770e", "uid": "40466951675f4545a54aab0ef4112bd5"}, {"from_node_channel": 0, "from_node_uid": "22424635fb4d4c1685225c044664770e", "to_node_channel": 0, "to_node_uid": "0f99856c3910415284cb0de1b441a525", "uid": "578200f648a84a2a9bcb4366c05fc576"}, {"from_node_channel": 0, "from_node_uid": "22424635fb4d4c1685225c044664770e", "to_node_channel": 1, "to_node_uid": "0f99856c3910415284cb0de1b441a525", "uid": "014e3278ca1046aab4320df66ba710d9"}, {"from_node_channel": 0, "from_node_uid": "0f99856c3910415284cb0de1b441a525", "to_node_channel": 0, "to_node_uid": "17454268ffff41a9a6c4636d618f1fde", "uid": "c5479c2575b248d79d3dd895eb9394d3"}, {"from_node_channel": 0, "from_node_uid": "17454268ffff41a9a6c4636d618f1fde", "to_node_channel": 0, "to_node_uid": "f3a6361a59a7466f918a9321780df393", "uid": "6fb0a844dd554cd29fae4588bbf75c1b"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.audio.AudioInput", "py/state": {"chunk_rate": 60, "device_index": null, "num_channels": 2}}, "numInputChannels": 0, "numOutputChannels": 2, "uid": "341b261ea3e546aa84152dea2e2b86d3"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "192.168.9.241", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "f3a6361a59a7466f918a9321780df393"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.ColorWheelEffect", "py/state": {"cycle_time": 30.0, "luminocity": 0.5, "num_pixels": 150, "offset": 0.0, "saturation": 1.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "57ad95eb504b4116875dca620ef0d6b3"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.ColorWheelEffect", "py/state": {"cycle_time": 15.0, "luminocity": 0.5, "num_pixels": 150, "offset": 0.0, "saturation": 1.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "1b660e63dc2e4556b8ec958ff1ce2c64"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.SpectrumEffect", "py/state": {"chunk_rate": 60, "col_blend": "lightenOnly", "fft_bins": 64, "fmax": 6000, "fs": 48000, "n_overlaps": 4, "num_pixels": 150}}, "numInputChannels": 3, "numOutputChannels": 1, "uid": "22424635fb4d4c1685225c044664770e"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.Append", "py/state": {"flip0": true, "flip1": false, "flip2": false, "flip3": false, "flip4": false, "flip5": false, "flip6": false, "flip7": false, "num_channels": 2}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "0f99856c3910415284cb0de1b441a525"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.AfterGlowEffect", "py/state": {"glow_time": 2.0}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "17454268ffff41a9a6c4636d618f1fde"}}], "recordTimings": true}}
+{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "c90cb5be463d4b268a6ab780c6fa8015", "to_node_channel": 0, "to_node_uid": "3a0809a0f53847b1a5ec0f482a821ac6", "uid": "128d4c916f3043b7ae0a9d0c993b9be5"}, {"from_node_channel": 0, "from_node_uid": "d3039145201542a380f7bdfd48dc5a20", "to_node_channel": 1, "to_node_uid": "3a0809a0f53847b1a5ec0f482a821ac6", "uid": "628be4bce50b4bfd823829ae217fd0a9"}, {"from_node_channel": 0, "from_node_uid": "5f297636d31944d2b90f28aacbfe0be9", "to_node_channel": 2, "to_node_uid": "3a0809a0f53847b1a5ec0f482a821ac6", "uid": "7df57c764d994baf8f239fccc9e2bd56"}, {"from_node_channel": 0, "from_node_uid": "3a0809a0f53847b1a5ec0f482a821ac6", "to_node_channel": 0, "to_node_uid": "4740257e2cdf44d3bd66637fb9a26e21", "uid": "20071b7f8c024abb9da98870538fe272"}, {"from_node_channel": 0, "from_node_uid": "3a0809a0f53847b1a5ec0f482a821ac6", "to_node_channel": 1, "to_node_uid": "4740257e2cdf44d3bd66637fb9a26e21", "uid": "f01ee197608a4969b27eb16bbc6997b4"}, {"from_node_channel": 0, "from_node_uid": "4740257e2cdf44d3bd66637fb9a26e21", "to_node_channel": 0, "to_node_uid": "de59fd0b3168464cb9404ac9d93c53c6", "uid": "0c319fe544784d65b2f477afdf3681ab"}, {"from_node_channel": 0, "from_node_uid": "de59fd0b3168464cb9404ac9d93c53c6", "to_node_channel": 0, "to_node_uid": "5cfc597adbf54482bf2723f6a80bdac4", "uid": "d8018f145a244a5eb0862f1ccb241a16"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.audio.AudioInput", "py/state": {"chunk_rate": 60, "device_index": null, "num_channels": 2}}, "numInputChannels": 0, "numOutputChannels": 2, "uid": "c90cb5be463d4b268a6ab780c6fa8015"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "127.0.0.1", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "5cfc597adbf54482bf2723f6a80bdac4"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.ColorWheelEffect", "py/state": {"cycle_time": 30.0, "luminocity": 0.5, "num_pixels": 150, "offset": 0.0, "saturation": 1.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "d3039145201542a380f7bdfd48dc5a20"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.ColorWheelEffect", "py/state": {"cycle_time": 15.0, "luminocity": 0.5, "num_pixels": 150, "offset": 0.0, "saturation": 1.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "5f297636d31944d2b90f28aacbfe0be9"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.audioreactive.SpectrumEffect", "py/state": {"chunk_rate": 60, "col_blend": "lightenOnly", "fft_bins": 64, "fmax": 6000, "fs": 44100, "n_overlaps": 4, "num_pixels": 150}}, "numInputChannels": 3, "numOutputChannels": 1, "uid": "3a0809a0f53847b1a5ec0f482a821ac6"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.Append", "py/state": {"flip0": true, "flip1": false, "flip2": false, "flip3": false, "flip4": false, "flip5": false, "flip6": false, "flip7": false, "num_channels": 2}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "4740257e2cdf44d3bd66637fb9a26e21"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.AfterGlowEffect", "py/state": {"glow_time": 2.0}}, "numInputChannels": 1, "numOutputChannels": 1, "uid": "de59fd0b3168464cb9404ac9d93c53c6"}}], "recordTimings": true}}
