+++ configs/swimming.json
@@ -1,1 +1,1 @@
-{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "3d165e077d6d4f7cbc1374edf6f9b665", "to_node_channel": 0, "to_node_uid": "98d7dbddd1b642ec93c2fdacc4a4c7ad", "uid": "0f48d6d4a996447d9805dfc37998dd99"}, {"from_node_channel": 0, "from_node_uid": "98d7dbddd1b642ec93c2fdacc4a4c7ad", "to_node_channel": 0, "to_node_uid": "90a28ebb337f47b1901460427b2f3757", "uid": "442e1730709242c8bc9ff1718ecb72d2"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "192.168.9.241", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "90a28ebb337f47b1901460427b2f3757"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.StaticRGBColorEffect", "py/state": {"b": 236.0, "g": 150.0, "num_pixels": 300, "r": 55.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "3d165e077d6d4f7cbc1374edf6f9b665"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.effects.SwimmingpoolEffect", "py/state": {"max_speed": 30, "num_pixels": 300, "num_waves": 30, "scale": 0.2, "wavespread_high": 70, "wavespread_low": 30}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "98d7dbddd1b642ec93c2fdacc4a4c7ad"}}], "recordTimings": true}}
+{"py/object": "audioled.filtergraph.FilterGraph", "py/state": {"connections": [{"from_node_channel": 0, "from_node_uid": "2cac73d9759b499a859f5e9fef392a3f", "to_node_channel": 0, "to_node_uid": "c1854434a8424f3192a4d349d658f70b", "uid": "5bbb3eca0b2e44c5b8152e34ae979666"}, {"from_node_channel": 0, "from_node_uid": "c1854434a8424f3192a4d349d658f70b", "to_node_channel": 0, "to_node_uid": "8bfeee15dddc434f84cd14fd24522eb3", "uid": "93494b27c46a460b89e977398b0abe6a"}], "nodes": [{"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.devices.LEDOutput", "py/state": {"controller": {"client": {"_ip": "127.0.0.1", "_long_connection": true, "_port": 7890, "_socket": null, "py/object": "audioled.opc.Client", "verbose": false}, "py/object": "audioled.devices.FadeCandy"}}}, "numInputChannels": 1, "numOutputChannels": 0, "uid": "8bfeee15dddc434f84cd14fd24522eb3"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.colors.StaticRGBColorEffect", "py/state": {"b": 236.0, "g": 150.0, "num_pixels": 300, "r": 55.0}}, "numInputChannels": 0, "numOutputChannels": 1, "uid": "2cac73d9759b499a859f5e9fef392a3f"}}, {"py/object": "audioled.filtergraph.Node", "py/state": {"effect": {"py/object": "audioled.generative.SwimmingpoolEffect", "py/state": {"max_speed": 30, "num_pixels": 300, "num_waves": 30, "scale": 0.2, "wavespread_high": 70, "wavespread_low": 30}}, "numInputChannels": 2, "numOutputChannels": 1, "uid": "c1854434a8424f3192a4d349d658f70b"}}], "recordTimings": true}}
